rss-fetcher/requirements.txt
# 数据处理和Excel操作 (核心依赖)
pandas>=1.5.0          # 数据分析，版本要求较高以支持现代功能
openpyxl>=3.1.0        # Excel文件读取
xlsxwriter>=3.1.0      # Excel文件写入和格式设置

# 网络请求和爬虫 (核心依赖)
requests>=2.28.0       # HTTP请求，版本要求较高以修复安全问题
httpx>=0.24.0          # 异步HTTP客户端，现代替代方案
beautifulsoup4>=4.11.0 # HTML解析，版本要求较高
lxml>=4.9.0            # XML/HTML解析，版本要求较高以修复安全问题
feedparser>=6.0.0      # RSS/Atom feed解析，用于主题书目模块


# 浏览器自动化 (核心依赖 - ISBN获取模块)
playwright>=1.40.0     # 浏览器自动化，用于豆瓣登录和ISBN获取

# 重试机制 (核心依赖)
retrying>=1.3.4        # 重试机制，版本要求较高

# 配置管理 (核心依赖)
python-dotenv>=1.0.0   # 环境变量管理
pyyaml>=6.0            # YAML配置文件解析

# 大模型调用 (可选依赖，有错误处理)
openai>=1.40.0         # OpenAI API客户端

# 时间处理 (核心依赖)
python-dateutil>=2.8.2 # 时间处理增强，版本要求较高
pytz>=2023.3          # 时区处理

# 数据验证 (可选依赖 - 部分代码引用)
numpy>=1.24.0          # 数值计算，版本要求较高以兼容现代pandas

# 向量数据库 (图书向量化模块)
chromadb>=0.4.22       # 向量数据库，用于语义检索
tqdm>=4.66.0           # 进度条显示（可选）
